assume mime-decode hook always produces output in the locale's encoding
Just because the attachment comes in as some specific encoding we don't
need or want to preserve that in the return value of mime-decode.
The user's mime-decode hook is probably just doing something like
running HTML through elinks, which will always produce a consistent
encoding. Let's assume it matches the locale we are already running as.

For a specific example, Outlook sends messages with
Content-Type: text/html; charset=us-ascii, but we don't expect the
mime-decode hook to produce ASCII-only output in that case (because any
non-ASCII characters in the mail will be ruined).